customer delete checks for active rental && machine delete rental check fix

+++ customer-microservice/src/main/java/hu/machinerental/customer_microservice/service/CustomerService.java
@@ -1,11 +1,15 @@
 package hu.machinerental.customer_microservice.service;
 
+import java.time.LocalDate;
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.UUID;
 
 import org.springframework.stereotype.Service;
 
+import hu.machinerental.customer_microservice.exception.CustomerHasActiveRentException;
+import hu.machinerental.customer_microservice.feignclients.rental.RentalClient;
+import hu.machinerental.customer_microservice.feignclients.rental.dto.Rental;
 import hu.machinerental.customer_microservice.model.dto.CreateCustomerReq;
 import hu.machinerental.customer_microservice.model.entity.Customer;
 import hu.machinerental.customer_microservice.repository.CustomerRepository;
@@ -16,7 +20,7 @@ import lombok.RequiredArgsConstructor;
 public class CustomerService {
 	
 	private final CustomerRepository customerRepo;
-	
+	private final RentalClient rentalClient;
 	
 	public List<Customer> findAll() {
 		return (List<Customer>) customerRepo.findAll();
@@ -44,7 +48,17 @@ public class CustomerService {
 	}
 
 	public void deleteById(UUID id) {
-		// TODO check for customer has no active machine rent
+		List<Rental> activeRentals = rentalClient.findAllFiltered(null, null, LocalDate.now(), null);
+		boolean hasRentalById = false;
+		
+		if(activeRentals != null && !activeRentals.isEmpty()) {
+			hasRentalById = activeRentals.stream().anyMatch(r -> id.equals(r.getCustomerId()));
+		}
+		
+		if(hasRentalById) {
+			throw new CustomerHasActiveRentException();
+		}
+		
 		customerRepo.deleteById(id);
 	}
